pyproject.toml
[tool.poetry]
name = "deltabase"
version = "1.0.0"
description = ""
authors = ["uname-n <git@defend.sh>"]
readme = "README.md"
packages = [{include = "deltabase"}]

[tool.poetry.dependencies]
python = "^3.11"
polars = "^1.6.0"
deltalake = "^0.19.0"
ipython = {version="^8.26.0", optional=true}
openai = {version="^1.43.0", optional=true}
pandas = {version="^2.2.2", optional=true}
simple-salesforce = {version="^1.12.6", optional=true}
google-cloud-bigquery = {version="^3.26.0", optional=true}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.31"
ipykernel = "^6.29.5"
pandas = "^2.2.2"

[tool.poetry.extras]
magic = ["ipython"]
ai = ["ipython", "openai"]
salesforce = ["pandas", "simple-salesforce"]
bigquery = ["google-cloud-bigquery"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
